fix(b300): promote paged-KV page ids to int64 to fix production IMA
Production crash after ~20min of load: CUDA illegal memory access
(Xid 31, FAULT_PDE ACCESS_TYPE_VIRT_READ). Root cause: physical page
ids were multiplied by the page stride in int32. The production KV pool
has 212,628 pages while the int32 overflow threshold is
2^31 / (64*1*256) = 131072 pages, so ~38% of page ids wrapped to
negative/garbage offsets and read unmapped addresses. Intermittent and
load-dependent: only triggers once the allocator places a request on
high page ids, which is why offline fuzz (compact ids) never caught it
and the page-range guard (page < num_pages) passed.

- flash_block_sparse_index_triton.py: promote physical_page to tl.int64
  before page*stride (covers the V load using the same variable).
- block_sparse_attn_triton.py: promote page to tl.int64 (covers v_slot).
- block_sparse_attn_cutedsl.py: promote the local_tile page coordinate
  to cutlass.Int64 so crd2idx computes the whole offset in 64-bit.
- prefill.py: FLASHPREFILL_PAGE_GUARD=1 host-side defensive net that
  validates page ids against num_pages, logs the offending
  (batch, token, page, kv_len) and clamps to survive. It cannot catch
  THIS bug (in-range large ids) but guards against real allocator bugs.

Verified with compute-sanitizer offline: previously faulting page
139000 and page 200000 (num_pages=205000, page*stride=3.28e9) now
produce 0 errors and finite outputs.

+++ flashprefill_ops/flashprefill/block_sparse_attn_cutedsl.py
@@ -520,7 +520,13 @@ class FlashPrefillCuteDSL(FlashAttentionForwardAmpere):
 
         # physical page for this logical tile (page_size == n_block_size, so the
         # page id is exactly the row-block index into the flattened KV cache).
-        page = p.mPT[p.batch, tile]
+        # int64: page ids index a large global KV pool (num_pages can exceed
+        # 2^31 / (page_size*num_kv_heads*head_dim) ~= 131072). Used raw as the
+        # local_tile coordinate, `page * row_stride` (== page*16384 here) is
+        # computed in int32 and overflows -> illegal address. Promote to int64
+        # so crd2idx does the whole offset in 64-bit. (Xid 31 fix; mirrors the
+        # triton kernels.)
+        page = cutlass.Int64(p.mPT[p.batch, tile])
 
         # --- load K page -> sK, V page -> sV (up-cast fp8->bf16 handled below) ---
         # mK/mV are flattened to (num_pages*page_size, num_kv_heads, head_dim);
